docx-to-excel-skill/requirements.txt
# Word到Excel转换技能依赖

# 核心依赖
python-docx>=0.8.11      # Word文档处理
openpyxl>=3.1.2          # Excel文件处理
pandas>=2.0.0            # 数据处理和分析

# 可选依赖（用于高级功能）
xlrd>=2.0.1              # 读取旧版Excel文件 (.xls)
xlwt>=1.3.0              # 写入旧版Excel文件 (.xls)
XlsxWriter>=3.1.0        # 高级Excel功能
python-pptx>=0.6.21      # PowerPoint文件处理（如果也需要处理PPT）

# 开发依赖
pytest>=7.4.0            # 测试框架
black>=23.0.0            # 代码格式化
flake8>=6.0.0            # 代码检查

# 安装命令
# pip install -r requirements.txt

# 最小安装（仅核心功能）
# pip install python-docx openpyxl pandas